Provide fallback pagination when Link header unavailable
Rather than relying on the Link headers to know if there are more pages
to load, we can just keep loading pages until they provide no additional
versions to parse. Using the Link headers is preferable, but this
provides an option to fall back on if the headers are not available.