Run on Node 20 (#66)
GitHub is no longer going to support running actions on Node 16 and requires upgrading to Node 20 (see blog). This PR addresses that with updates to:

 * action runtime
 * workflows
 * node types

Tests passed when run locally. Closes #65.

Add @types/node to devDependencies to ensure we get the proper version and bump @vercel/ncc to the latest, and update package-lock using `npm install` which also prepares the package.