Support for aarch64/arm64 ubuntu-24.04-arm runner
Add support for aarch64 Linux installs of cmake, and add CI test
coverage for the ubuntu-24.04-arm runner in particular.

To implement this change, I changed a bit how universal binaries are
handled internally on Darwin.  On Mac runs, packages marked with the
`universal` tag are presumed compatible.  If a universal package is not
available, then `x86_64` packages are used.  This behavior should
support both x86_64 and arm runners for all currently tested releases of
cmake.